refactor #2, still doesn't work, also define a path planner interface
