+++ src/thunderbots/software/ai/navigator/path_planner/path_planner.h
@@ -1,0 +1,11 @@
+#pragma once
+#include <vector>
+#include <optional>
+#include "geom/point.h"
+
+class PathPlanner {
+public:
+    virtual std::optional<std::vector<Point>>
+        findPath(const Point& start, const Point& dest) = 0;
+    virtual ~PathPlanner() = 0;
+};
